requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
celery==5.6.2
redis==7.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Pillow==11.3.0
django-filter==25.2
pyyaml==6.0.3